Auth0 client secret is not base64 encoded anymore.
The client secrets provided for Auth0 clients used to be base64 encoded
but aren't anymore [0].

Revision of dgrijalva/jwt-go currently used, expects the secret that is
provided to `jwt.Parse` to be encoded.

For now, encode the secret pulled from the env variable before using it
to register the Application.

[0] https://community.auth0.com/questions/3666/invalid-signature-error-on-jwtio-and-in-backend